# docs/swagger.yaml
definitions:
  api.LoginInfo:
    properties:
      code:
        type: string
      codeId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - code
    - codeId
    - password
    - username
    type: object
  api.OrgAdd:
    properties:
      address:
        type: string
      createdAt:
        type: string
      creditCode:
        type: string
      id:
        example: "0"
        type: string
      menuIds:
        items:
          type: string
        type: array
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  domain.Menu:
    properties:
      createdAt:
        type: string
      icon:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      path:
        type: string
      pid:
        example: "0"
        type: string
      sort:
        type: integer
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  title: Admin API
  version: "1.0"
paths:
  /login/account:
    post:
      consumes:
      - application/json
      parameters:
      - description: login info 信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.LoginInfo'
      produces:
      - application/json
      responses: {}
      summary: login 用户登录
      tags:
      - login 登录
...
  /menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: name 名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: List Menus 获取菜单列表
      tags:
      - menus 菜单
    post:
      consumes:
      - application/json
      parameters:
      - description: Menu info 菜单信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.Menu'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create menu 创建菜单
      tags:
      - menus 菜单
  /menus/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete menu 删除菜单
      tags:
      - menus 菜单
    get:
      consumes:
      - application/json
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Menu 获取菜单
      tags:
      - menus 菜单
    put:
      consumes:
      - application/json
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Menu info 菜单信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.Menu'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update menu 更新菜单
      tags:
      - menus 菜单
  /orgs:
    get:
      consumes:
      - application/json
      parameters:
      - description: name 名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: List orgs 获取机构列表
      tags:
      - orgs 机构
    post:
      consumes:
      - application/json
      parameters:
      - description: Org info 机构信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.OrgAdd'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create org 创建机构
      tags:
      - orgs 机构
  /orgs/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete org 删除机构
      tags:
      - orgs 机构
    get:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Org 获取机构
      tags:
      - orgs 机构
    put:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      - description: Org info 机构信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.OrgAdd'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update org 更新机构
      tags:
      - orgs 机构
  /orgs/{id}/menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Org menus 获取机构菜单
      tags:
      - orgs 机构
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"